requirements.txt
llama-cpp-python
torch
